+++ src/components/Tab/Searchbar.jsx
@@ -1,48 +1,75 @@
-import React, { Component } from 'react'
+import React, { Component } from 'react';
 
 export default class SearchBar extends Component {
     constructor(props) {
-        super(props)
+        super(props);
         this.state = {
             value: 'https://www.google.com/search?q=',
             isFocusing: false
-        }
-    }
-
-    handleChange = (e) => {
-        this.setState({ value: e.target.value })
-    }
-
-    handleSubmit = (e) => {
-        e.preventDefault()
-        this.setState({ value: e.target.value })
-        // this.props.onSubmit(this.state.value)
-    }
-
-    parseURL(url) {
-        const urlObj = new URL(url);
-        return {
-            protocol: urlObj.protocol,
-            domain: urlObj.hostname,
-            path: urlObj.pathname,
-            raw: url
         };
     }
 
-    handleFocus = (e) => {
-        this.setState({ isFocusing: true })
-        e.target.select()
+    handleChange = e => {
+        this.setState({ value: e.target.value });
+    };
+
+    handleSubmit = e => {
+        e.preventDefault();
+        this.setState({ value: e.target.value });
+        // this.props.onSubmit(this.state.value)
+    };
+
+    parseURL(url) {
+        const urlObj = new URL(url);
+        return urlObj;
     }
 
-    handleBlur = (e) => {
-        this.setState({ isFocusing: false })
-    }
+    handleFocus = async e => {
+        this.setState({ isFocusing: true });
+        await e.target.select();
+    };
+
+    handleBlur = e => {
+        this.setState({ isFocusing: false });
+    };
 
     render() {
         return (
-            <div className='searchBar'>
-                <input type='text' value={this.state.isFocusing ? this.state.value : this.parseURL(this.state.value).domain} onChange={this.handleChange} onFocus={this.handleFocus} onBlur={this.handleBlur} />
+            <div className='search-bar'>
+                {/* Have to open a modal that display information about the security of the website */}
+                <button className='sslProtocol'>
+                    {this.parseURL(this.state.value).protocol.startsWith('https:') ? (
+                        <svg width="50px" height="50px" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                            <path fillRule="evenodd" clipRule="evenodd" d="M5.5 10V7C5.5 5.27609 6.18482 3.62279 7.40381 2.40381C8.62279 1.18482 10.2761 0.5 12 0.5C13.7239 0.5 15.3772 1.18482 16.5962 2.40381C17.8152 3.62279 18.5 5.27609 18.5 7V10H19C20.6569 10 22 11.3431 22 13V20C22 21.6569 20.6569 23 19 23H5C3.34315 23 2 21.6569 2 20V13C2 11.3431 3.34315 10 5 10H5.5ZM9.52513 4.52513C10.1815 3.86875 11.0717 3.5 12 3.5C12.9283 3.5 13.8185 3.86875 14.4749 4.52513C15.1313 5.1815 15.5 6.07174 15.5 7V10H8.5V7C8.5 6.07174 8.86875 5.1815 9.52513 4.52513Z" fill="#000000" />
+                        </svg>
+                    ) : (
+                        <svg width="50px" height="50px" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                            <line fill="none" stroke="#000000" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" x1="12" x2="12" y1="8" y2="12" />
+                            <line fill="none" stroke="#000000" stroke-linecap="round" stroke-linejoin="round" stroke-width="2" x1="12" x2="12" y1="16" y2="16" />
+                            <circle cx="12" cy="12" data-name="--Circle" fill="none" id="_--Circle" r="10" stroke="#000000" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" />
+                        </svg>
+                    )}
+                </button>
+                <input
+                    type='text'
+                    className='search-input'
+                    value={
+                        this.state.isFocusing ? (
+                            this.state.value
+                        ) : (
+                            this.parseURL(this.state.value).hostname
+                        )
+                    }
+                    onChange={this.handleChange}
+                    onFocus={this.handleFocus}
+                    onBlur={this.handleBlur}
+                />
+                <button className='refreshButton'>
+                    <svg width="50px" height="50px" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                        <path d="M4,12a1,1,0,0,1-2,0A9.983,9.983,0,0,1,18.242,4.206V2.758a1,1,0,1,1,2,0v4a1,1,0,0,1-1,1h-4a1,1,0,0,1,0-2h1.743A7.986,7.986,0,0,0,4,12Zm17-1a1,1,0,0,0-1,1A7.986,7.986,0,0,1,7.015,18.242H8.757a1,1,0,1,0,0-2h-4a1,1,0,0,0-1,1v4a1,1,0,0,0,2,0V19.794A9.984,9.984,0,0,0,22,12,1,1,0,0,0,21,11Z" />
+                    </svg>
+                </button>
             </div>
-        )
+        );
     }
 }