+++ src/components/Buttons/AddBookmarkButton.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { useDispatch } from 'react-redux';
+import { AddTabIcon } from '../UI/Icons';
+
+export default function AddBookmarkButton() {
+    const dispatch = useDispatch();
+
+
+    return (
+        <button className='addBookmarkButton'>
+            <AddTabIcon />
+        </button>
+    )
+}