+++ src/components/Webview/Webview.jsx
@@ -1,15 +1,26 @@
-import React, { useRef } from 'react';
+import React, { useRef, useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { addWebview, toggleWebview } from '../../actions/tabs.actions';
 
-const WebView = ({ tabId }) => {
+const WebView = ({ tabId, defaultURL }) => {
     const webviewRef = useRef(null);
+    const dispatch = useDispatch();
+
+    useEffect(() => {
+        webviewRef.current.addEventListener('dom-ready', () => {
+            dispatch(addWebview(tabId, webviewRef.current));
+            dispatch(toggleWebview(tabId, true));
+        }, { once: true });
+    }, [dispatch, tabId]);
 
     return (
         <webview
             ref={webviewRef}
-            src={"https://www.google.com/"}
+            src={defaultURL}
             className='webview'
             plugins="true"
             allowpopups="true"
+            key={tabId}
         ></webview>
     );
 };