+++ public/constants.js
@@ -1,25 +1,9 @@
 module.exports = {
     channels: {
-        //App
-        CLOSE_APP: 'closeApp',
-        //Tabs
-        RELOAD_TAB: 'reloadTab',
-        BACK_IN_TAB: 'backInTab',
-        PREVIOUS_IN_TAB: 'previousInTab',
-        CLOSE_TAB: 'closeTab',
-        //Webview
-        SEARCH_IN_TAB: 'searchInTab',
-        OPEN_DEVTOOLS: 'openDevTools',
-        OPEN_URL_IN_NEW_TAB: 'openUrlInNewTab',
-        OPEN_URL: 'openURL',
-        OPEN_URL_IN_RENDERER: 'openURLInRenderer',
-        //UserPreferences
-        GET_USER_PREFERENCES: 'getUserPreferences',
         //Bookmarks
-        GET_BOOKMARKS: 'getBookmarks',
-        ADD_BOOKMARK: 'addBookmark',
-        REMOVE_BOOKMARK: 'removeBookmark',
-        //Tabs
-        GET_TABS: 'getTabs',
+        ADD_BOOKMARK: 'ADD_BOOKMARK',
+        REMOVE_BOOKMARK: 'REMOVE_BOOKMARK',
+        UPDATE_BOOKMARK: 'UPDATE_BOOKMARK',
+        GET_BOOKMARKS: 'GET_BOOKMARKS',
     },
 };