+++ public/ipc/bookmarks/addBookmark.js
@@ -1,0 +1,10 @@
+const { channels } = require("../../constants");
+const { bookmarks } = require("../../storedData");
+
+module.exports = {
+    name: channels.ADD_BOOKMARK,
+    once: false,
+    async execute(webContents, _, data) {
+        bookmarks.set('bookmarks', [...bookmarks.get('bookmarks'), data]);
+    }
+}