+++ changelog.md
@@ -44,7 +44,6 @@
         - New SearchBar based on Safari
         - New TabManager
         - Bookmark Menu at the left
-        
         In PR #6 
         - Bookmark Menu Button at the right with arrow down
         - Add Default Themes
@@ -53,7 +52,7 @@
         - Add Tab Reducer and Actions
         - Deconstruct The TabManager file to TabBar.jsx, TopBar.jsx and RouterManager.jsx
         - Add Close button in TopBar
-
         In PR #7 
         - Add Icons.jsx
-        <!-- - Add Transition to display the closse button in TabBar -->
+        In PR #8
+        - Add Transition to display the closse button in TabBar